Added sulfur padding to combine_xyzs.py
This runs at the end of 'track_molecules' task for molfind, finds the max # of atoms in any frame of the output xyzs, orders the alanine atoms first and other fragments after, with sulfur pad atoms at the bottom of the list.
Sulfurs are added at the centroid over every frame, to be used as an anchor point for the visualization.